TestMyInternetSpeed/requirements.txt
speedtest-cli